Rewrite clustering logic from kmeans approach

+++ HiveWays.FleetIntegration/Business/ClusterConfiguration.cs
@@ -2,7 +2,7 @@
 
 public class ClusterConfiguration
 {
-    public int ClusterRadius { get; set; }
+    public decimal ClusterRadius { get; set; }
     public int VehicleLength { get; set; }
     public int VehicleWidth { get; set; }
     public int VehicleDistance { get; set; }
